chore: Sovereign Hardening v7.1.18 - Security & Performance Remediation (Batch 1-7)
- Enforced strict multi-tenancy via HasHouseholdScope (CLI bypass removed)
- Implemented transactional row-level locking (lockForUpdate) in financial actions
- Hardened business logic with status-based validations (Holidays/Goals)
- Optimized performance with N+1 fixes and caching strategies
- Enhanced security headers and sudo mode device binding
- Resolved static analysis (PHPStan/Rector/Insights) and precision warnings
- Verified with comprehensive test suite (27+ feature tests)

+++ apps/backend/app/Http/Resources/WealthHistoryResource.php
@@ -34,7 +34,7 @@ class WealthHistoryResource extends JsonResource
 
         return [
             'month' => $date instanceof Carbon ? $date->format('M') : 'N/A',
-            'value' => (int) $this->total_value,
+            'value' => (float) $this->total_value,
             'year' => (int) $this->year,
             'raw_month' => (int) $this->month,
         ];
